key: implement Server shutdown
This commit implements a mechanism for calling
code to shutdown a running Server using Context.

The ServeTCP method now accepts a context that,
when `<-ctx.Done()` returns, 1st tries a graceful
shutdown via `http.Server.Shutdown`. If this graceful
shutdown takes longer than the provided graceful shutdown
timeout, the server gets closed immediately.

For instance, calling code can now listen for OS signals
(e.g. SigTerm) and, if one is encountered, then call the
CancelFunc of the context passed to ServeTCP.
In particular, this is done by the server command in cmd/key.