improve AES key-derivation
This commit changes the AES key derivation function
used for AES-GCM keys. In general, we reuse the
key derivation of AES-GCM-SIV (RFC 8452). However,
AES-GCM-SIV derives a key from a 96 bit nonce while
we derive a key from a 128 IV / nonce.

So we modify the AES-GCM-SIV key derivation:
```
  k0 = E(k, 0 || nonce)[0..8]
  k1 = E(k, 1 || nonce)[0..8]
  k2 = E(k, 2 || nonce)[0..8]
  k3 = E(k, 3 || nonce)[0..8]
```
as following:
```
  t0 = E(k, 0 || iv[0..12])
  t1 = E(k, 1 || iv[4..16])
  t2 = E(k, 2 || iv[0..12])
  t3 = E(k, 3 || iv[4..16])

  k0 = t0[0..8]  ^ t1[0..8]
  k1 = t0[8..16] ^ t1[8..16]
  k2 = t2[0..8]  ^ t3[0..8]
  k3 = t2[8..16] ^ t3[8..16]
```

The idea here is that each 64 bit block of
the derived key depends on the entire IV.
Therefore, we XOR the PRP (AES) outputs when
the inputs are different due to the IV.