fs: add new file system key store
This commit adds a new key store for saving
secret keys to the file system.

A key server can be configured to store key
files in a directory. Therefore, one adds
```
[fs]
path = "<path-to-dir>"
```
to the key server configuration file. The server
will then create that directory (if it does not
exist). Whenever a new key is created the server
will store it as file inside that directory.
Similarly, when a key is requested for the first
time the server will load it from the file inside
the directory.

A secret key file has the same name as the secret
key. For instance the secret key `my_key` is stored
in a file `my_key`.