cache: new internal package cache
This commit adds a new internal package `cache`
that implements an in-memory cache for secret keys.

It implements common and shared logic between multiple
key store implementations (FS, Vault, ...).

Further, the cache supports garbage collection / expiry of
(not frequently used) cache entries. The main idea behind
this functionality is that when a cache entry expires the
key server has to fetch it from the storage backend again.
By controlling how often / when the key server has to reach
out to the backend certain security guarantees can be enforced.
For instance, it can be guaranteed that a secret key never survives
longer than X when deleted from the backend storage.