server: add CLI flag controlling client cert verification
This commit adds the `--mtls-auth` cli flag to the server
command. Now, the server can be configured to either
require & verify (default) or require but don't verify the
client certificate.

In theory, we could also allow a third option such that
the server also accepts client connections if the client
does not present a certificate at all. However, there is
no point in doing so since we cannot compute an identity
from a request without a TLS certificate -> All requests
without a certificate will be denied by the policy system.