make CLI flag parsing smarter
This commit makes the CLI flag parsing for commands smarter.
In particular, it now doesn't matter whether you pass a flag
between the command and its args, between two args or after the
args.

Further, this commit renames the flag that causes the client to
skip X.509 certificate verification to `k` or `--insecure` (like curl).