add new sub-command: tool
This commit adds the new `tool` sub-command and moves
the `identify` command to `tool`.

The purpose of the `tool` sub-command is to collect
useful commands that don't require interacting with
a key server and/or make configuration simpler.

In particular, there is now an `identity` sub-command
that provides identity management tools:
 - `key tool identity of <certificate>`
 - `key tool identity new [opt] <name>`

The `of` command is the replacement for `identify`.
The `new` command creates a private key and certificate
with the given name as common-name. It makes creating
(tests) identities more convenient.