add `--mlock` flag to lock memory pages
This commit adds the `--mlock` flag to
the server command. When specified mlock
ensures that the OS will not swap memory
packages of the key server to the disk.

This prevents leaking secrets by accident
through OS swapping. However, this is only
available on linux (for now, unix systems
in general support it).

On windows there would be the `VirtualLock`
syscall but it can only lock a specific
memory range, not all pages.

For further information see: `man mlockall`