use error values for CLI commands
This commit replaces the error handling
of CLI commands. Before, when a command
encountered an error it would call `failf`
which would invoke `os.Exit` to terminate
the program.

Now, a command returns an error such that
it becomes more expressive when we abort the
execution - in particular in `if` blocks.