add support for logging error events (error log)
This commit adds a error log system. In general,
the server distinguishes between error and audit
events.
In particular, an error event is logged when e.g.
the server fails to reach an external key store
b/c of some network problem or encounters a
malformed secret.

In contrast, an audit event is an interaction between
any client and the server - for instance trying
to create a key or assign a policy to an identity.
These events are separated from error events and
deserve a separate log system.

Further, this commit fixes a bug in the AWS Secrets
Manager. Before this commit, the kes server would
not honor cache expiry deadlines for the Secrets
Manager key store.