fix incorrect identity assignment during startup (#10)
This commit fixes a bug in the identity<->policy
assignment during the server startup.

Now, the server does not assign a policy to the
unknown identity (empty string). Instead, such
policy<->identity mappings are ignored.

The previous behavior was not correct but also
not a security issue since the unknown identity
is always rejected by the server when doing a policy
lookup.