improve API path validation (#9)
This commit improves the validation of
the request URL path. Now, the `ValidatePath`
handler function accepts a API path pattern
and tries to match the request URL path against
this pattern.
Only if the URL path matches that pattern the
request gets handled. Otherwise, we return 400
(bad request) to the client.

This simplifies adding new (more complex) APIs and
makes it easier to think about security implications
(See: directory traversal a.o.).