add basic support for audit logging (#7)
This commit adds basic support for logging
audit events to files.

Therefore, it adds a new `SystemLog` type
that allows attaching and detaching log targets
(files, sockets, http connections, etc.) to
a `log.Logger`.

Now, we can dynamically add and remove log targets
to the currently configured logger. This is in service
for dynamically inspecting and at/detaching to logging -
which will be added in the future. This commit paves
the path for this.

The audit log itself is just a `http.HandlerFunc`
that extracts some "interesting" request parameters
(API path, X.509 identity, ...) and logs them together
with the response before writing back to the client.

At the moment the log contains the current time,
request API path, client identity and response status
code. However, this can be extended in the future.
So, the current audit log implementation is just a MVP.