update cache expiry default values
This commit adjusts the cache expiry
default values to:
 - discard all entries after `5 min`
 - discard entries that haven't been
   used for more then `20 sec`

The value for `all` is reduced significantly
since:
```
1 Month / 5 min = 30 * 24 * 60 / 5 = 8640
```
So, by discarding all cache entries we make 8640
requests to the key store. For ref: AWS charges
per 10,000 requests (e.g. 10,000 req = 0.03 ct).
This means that the KES server has to authenticate
itself to the key store and request access to the
master/secret key every 5 min - which is quite
conservative.

Further, if an application is not using the
master/secret key for more 20 sec the KES
server will discard it as well. This is again
a conservative choice. However, for high-performance
applications 10 sec or less may also be appropriate.

Now, the expiry values are quite conservative and
should provide a reasonable security level.