add version information to binary and server (#17)
This commit adds version information to the
binary and the server.

Now, when making a release the commit tag
becomes the release version. So, running
```
cd ./cmd/kes && ./release.sh
```
after tagging a new release (e.g. v1.0.0)
will include `v1.0.0` as version into the binary.
If HEAD has no tag the release script will print
a warning and use `0.0.0-<commit-ID>` as release
version.

The binary version is printed when running:
```
kes --version
```

Further, the server handles GET requests
to `/version` and respond with a JSON object
containing the version of the server.

Note that the `/version` route is available to
anyone who can authenticate (via mTLS) to the
server. There is no policy that allows/denies
access to `/version`.