add support for TLS proxies
This commit adds support for TLS proxies such that
there can be 0, 1 or multiple TLS proxies between
one (or more) kes clients and a kes server.

The identity of each proxies (direct neighbor of the kes server)
must be added to `proxy.identities`.
Further, the proxy which is a direct neighbor of the kes client
must forward the kes client certificate as part of the request
header.

So, in the most simple (and most common) case there is one TLS
proxy - e.g. a nginx load balancer - between the kes server and
kes client. This nginx has to extract and forward the client
certificate (request headers) and has to authenticate itself
via its own nginx client certificate to kes as proxy.

Optionally, the nginx may verify the kes client certificate as well.

If there would be two load balancers then the one closer to the
client has to extract the forward the client certificate while
the one closer to the kes server has to authenticate itself to the
kes server. This must be the case for any number ( >= 2) TLS proxies.

As an example, the following nginx configuration shows how to
setup a demo load balancer on your local machine:
```
http {
   server {
     listen              443 ssl;
     server_name         localhost;

     # Avoid buffering the response. Especially for audit log tracing.
     proxy_buffering off;

     # The nginx private key and certificate presented to clients
     # connecting to it.
     ssl_certificate     /home/andreas/nginx.crt;
     ssl_certificate_key /home/andreas/nginx.key;

     # Require a client certificate but don't verify it.
     # See nginx docs for more details.
     ssl_verify_client   optional_no_ca;

     location / {
        # KES server endpoint
        proxy_pass                  https://localhost:7373;

        # Enforce TLSv1.3 - KES supports it.
        proxy_ssl_protocols         TLSv1.3;

        # This demo config assumes the KES server uses
        # self-signed certificates.
        proxy_ssl_verify            off;

        # The private key and client certificate used by nginx
        # to authenticate itself to the kes server
        # The output of: 'kes tool identity of <nginx-client.cert>'
        # must be added to the proxy.identities section in the
        # kes config file.
        proxy_ssl_certificate       <nginx-client.cert>;
        proxy_ssl_certificate_key   <nginx-client.key>;

        # The header used by nginx to forward the escaped and
        # encoded certificate of the KES client. This value
        # must match the entry in the tls.proxy.header section
        # in the KES config file.
        proxy_set_header X-Tls-Client-Cert $ssl_client_escaped_cert;

        # We need keep-alive for audit log tracing
        proxy_http_version 1.1;
     }
  }
}
```

Fixes #18