avoid calling ResponseWriter.WriteHeader twice (#20)
This commit fixes a bug in the `AuditResponseWriter`
in combination with the /v1/log/audit/trace API.

The bug is caused by calling the `ResponseWriter.WriteHeader`
method twice.
This happens when a client hits the /v1/log/audit/trace
API which adds the `http.ResponseWriter` (which is a
`AuditResponseWriter`) to the list of log output dest.

Now, when the handler function tries to write 200 OK
to the client it calls `WriteHeader` - which sends the
audit log event to all destinations (including the client
which triggers an implicit 200 OK). Then the `WriteHeader`
method actually tries write a status code to the client -
which has been done already by the logger.

This commit fixes this by having two state variables which
ensures that as soon as a implicit write happens no further
`WriteHeader` calls are performed.