implement encrypted secrets using a KMS
This commit implements encrypted secrets.
Before, KES stored all secrets as plaintext at the
backend secret store. Therefore, KES assumed/required
that the secret store is a secure/trusted component.

While this assumption may be true for secret stores like
Vault or AWS SecretsManager it is not true in the general
case. Further, even if the secret store is considered trusted
it may be desirable to separate the trust between the secret
storage provider and another party.

Therefore, a KES server may want to encrypt secrets before
storing them at the secret store. In general, KES encrypts
secrets with an (external) KMS - like Vault, AWS-KMS or
Azure Vault. Therefore, KES sends a secret to the KMS before
writing it to the secret store. However, it doesn't send the
secret as plaintext to the KMS since that would reveal the
secret to the operator of the KMS / anyone who can inspect
KMS operations.

Instead, it generates a random (local) cryptographic key
when creating a secret and encrypts the secret with that key
*before* sending the secret to the KMS. So, the KMS only
sees ciphertexts (`Enc(local_key, secret)`). The local key
is then embedded into the value written to the secret store.

This construction ensures that neither the secret store provider
nor the KMS provider can decrypt an encrypted secret on their own.
So, it is possible to leverage e.g. a cloud KMS provider (AWS, Google,
MS, ...) to encrypt secrets on some local storage without sharing
these secrets with the cloud provider.

***

Therefore, this commit changes the architecture of the exiting code
quite a bit:

There is now a local store (`secret.Store`) and a remote store
(`secret.Remote`). The `secret.Store` implements en/decryption
(via the `secret.KMS` interface) and is responsible for caching
secrets.
The `secret.Remote` interface defines what operations a secret
store has to provide and must be implemented by all secret store
implementations.

Since, a secret store now can contain plaintext and encrypted
secrets the `secret.Remote` consumes strings while `secret.Store`
consumes `secret.Secret`.
However, for now we demand that the secret store returns only
plaintext secrets or encrypted secrets - but not both. When
the KES server is configured to use a KMS it will always try
to decrypt values from the secret store - including plaintext
secrets which should fail.

An encrypted secret has the following format:
```
{
  "local_key": {
    "bytes": "4BIYQkplmutXWkMg38enraePnsYQauIQkhzBhD7xs+U="
  },
  "kms_key": "5f91e7d9-a376-47cd-ad6f-3485abf9122c",
  "ciphertext":"<base64-encoded-ciphertext>"
}
```

The `local_key` is the cryptographic key used to encrypt
the plaintext secret before sending it to the KMS.
The `kms_key` is the name of the cryptographic key at
the KMS used encrypt the secret (ciphertext).
So, the result of encrypting the plaintext first with
the local key and then with the KMS is `ciphertext`.

***

Apart from this conceptual change, this commit also
adds a `secret.KMS` implementation for the AWS-KMS and
extends the KES server configuration to provide an
AWS-KMS configuration.