ignore empty static credentials
This commit fixes a bug affecting the AWS SecretsManager
key store.
When no access key, secret key or session token is provided
the credentials passed to the AWS SDK should be `nil`. This
causes the AWS SDK to look for credentials in AWS env. variables,
or EC2 metadata.

I've verified that this commit fixes the bug reported in #26 by
creating an AWS EC2 instance and specifying no credentials in the
config file.
If the correct role (e.g. SecretsManager Read/Write) is attached to
the instance specifying no credentials causes the kes server (via AWS
SDK) to fetch the temp. credentials from the EC2 metadata.

Fixes #26