add install instructions to README.md
This commit adds how to install KES instructions
to the README.md. This is in particularly necessary
since we now support signed releases and the public key
for verification must be presented to users in way that
ensures that it is not modified by an attacker.

With this change the public key is show as part of the README.md
which is part of the git repository. So only people with push/commit
access can change it and if they do the change will be preserved
in the git history (or they have to rewrite the history which most
probably will be detected).

Further, this commit adds docker install instructions.