stabilize `kes.Client` and `New*` functions
This commit stabilizes the KES client API.
The `Client` type now exports the KES server
endpoint and the HTTP client.

This is a very flexible API in the sense that
a user can configure any custom transport
configuration (timeouts etc.) or implementation
(TCP-TLS, QUIC, ...).
So, this way the `Client` type can be used even
if the underlying protocol layers will change.

Further, there are two convenience functions:
 - `NewClient`
 - `NewClientWithConfig`

which will be sufficient for most scenarios
where the user just provides a client certificate
or a TLS configuration.