stabilize the client key API
This commit stabilizes the client key API.
Currently, key API consists of the five operations:
 - `CreateKey`
 - `DeleteKey`
 - `ImportKey`
 - `GenerateKey`
 - `Decrypt`

`CreateKey` and `DeleteKey` create resp. delete
a cryptographic key at the KES server. `ImportKey`
behaves as `CreateKey` but the client specifies the
cryptographic key instead of the server generating one.

The `GenerateKey` API requests a new data encryption key
(`DEK`) from the server. In general, the DEK can be used
for more than just encryption - e.g. compute a MAC - but
the term data encryption key and abbreviation `DEK` is
common for such keys in the context of key-management.

The DEK has a plaintext and a ciphertext representation.
The plaintext should be used for cryptographic operations
while the ciphertext should be stored at a durable location,
does not need to be secret and can be decrypted by the server
whenever the client needs the plaintext representation.

The `Decrypt` takes a ciphertext and an optional context and
asks the server to decrypt the ciphertext with the named crypto.
key.
In general, `Decrypt` is not limited to just decrypting a DEK's
ciphertext. It is more generic and can be used to decrypt any
ciphertext blob produced by the server. Generating a DEK can be
seen as a special case of an `Encrypt` API where the server
also chooses the plaintext instead of the plaintext providing it.

At the moment, the server does not provide an `encrypt` API.
However, that may change in the future.

Currently the key-API methods don't take any optional arguments.
In the future there may be a demand to extend these APIs.
For example users may want to e.g. create not just symmetric
keys but also e.g. an ECDSA / Ed25519 key.

It would have been possible to provide an API with optional arguments,
e.g.:
```
func (c *Client) CreateKey(name string, opts ...CreateOptions)
```

I decided to not go down this route right now to make the API reflect
the most common / simplest case. If there is a need to extend the API
I propose the following pattern:
```
func (c *Client) CreateKey(name string)

func (c *Client) CreateKeyWith(name string, opts ...CreateOptions)
```

This way there is a simple API that works in 90% of the cases and one
(slightly) more complex API that can address the remaining 10%.