drop encrypted secrets (KMS) support and revert etcd
This commit removes support for encrypted secrets
at the key store as well as the etcd key store
added by commit: 71ba6d2.

This is a breaking change. The next version will not
accept an existing KMS configuration. Further, this
commit removes the 'insecure' AES key derivation fixed
in v0.6.1. Keys generated before this version cannot
be decrypted with the any future version.

This change is made with the idea to remove any
non-necessary complexity and instead optimize
for the 90% use case.

The KMS<->Key Store separation provides better
flexibility since some KMS implementations don't
provide a secure key store (K/V store). For
example AWS has separated the KMS (AWS-KMS) from
a secure key/value implementation (AWS SecretsManager).

The concept of the KMS being separate from the key store
has been introduced to address this limitation.
However, it has turned out that this is quite confusing
for people that are new to the project. Therefore, we
decided to remove the KMS (that protects the key store)
completely and demand that the key store has be secure.

This means that the key store has to be implemented by a
KMS that supports an encrypted key/value store.

```
 ┌────────────┐        ┌────────────┐      ┌───────┐
 │ KES Client ├────────┤ KES Server ├──────┤  KMS  │
 └────────────┘        └────────────┘      └───────┘
```

It has turned out that this concept is as powerful as the
previous abstract - even though being simpler to understand
from an architectural perspective.

It seems like the above design excludes any non-secure
Key Store but this is actually not the case.

An insecure key store (e.g. etcd) can be turned into an
encrypted one via a KMS-adapter that uses a "feature-limited"
KMS (e.g. AWS-KMS) and encrypt the keys before writing them to
the key store.

```
 ┌─────────────┐        ┌───────────┐
 │ KMS Adapter ├─┬───┬──┤ Key Store |
 └─────────────┘ |   |  └───────────┘
                 |   |
                ┌┴───┴┐
                | KMS |
                └─────┘
```

The advantage of this approach is that 90% of the users won't have
to worry about the KMS/key store separation while the remaining 10%
can still use a key store that is separated from the KMS - via such
a KMS adapter.

This commit does not add a generic KMS adapter interface (e.g. gRPC
or REST). This will be added in the future.