restrict server API to HTTP2
This commit restricts the KES server API to
HTTP2.

The reason for this restriction is that future
API functionality probably will require bidirectional
streams between the server and the client.

In particular, the KES server will expose APIs that
return a response to the client even though the client
still sends more data. (streaming)

This is not possible with HTTP/1.0 and HTTP/1.1 since
the server must not return a response body to the client
before the server hasn't read the entire client request.

While, in theory, it would be possible to restrict only
those APIs to HTTP2, it seems better to enforce this
restriction for all APIs from the beginning. This avoids
mental complexity for clients - since they don't have to
remember which APIs are restricted and which are not.

Also, we can relax the restriction in the future much more
easily then putting additional limitations in place. The
later would break clients.