remove the restriction of only supporting HTTP/2
This commit allows clients to connect to the KES server
via any HTTP protocol (again). This commit basically reverts
b74bb38.

The reason for restriction to HTTP/2-only was the potential
addition of stream APIs for encryption and decryption.
However, nginx does not support HTTP/2 for backend connections.
See: https://trac.nginx.org/nginx/ticket/923

Therefore, it would not be possible to run a KES server behind
a nginx proxy since the nginx would not be able to connect to
KES.

Further, this commit fixes 3 typos in the reference config file.