implement request retry mechanism for KES client
This commit implements a request retry mechanism.
Now, client automatically retries a request if
the failure seems to be temporary.

In particular, if e.g. a KES server dies for some
reason before it has sent a response then the client
fails with an `*url.Error` wrapping an `io.EOF` or
an `io.ErrUnexpected`. In this case we retry the request
hoping that the KES server gets (automatically) restarted.