add /v1/key/encrypt server API
This commit adds a new server API for encrypting
(small) data with a master key. The server now
exposes a new API endpoint:
```
/v1/key/encrypt/<key-name>
```

It expects a JSON as request body:
```
{
  "plaintext": "...",
  "context": "..."     // optional
}
```
The plaintext is encrypted and authenticated with the master key.
More precisely, a "key encryption key" is derived from a random
IV: `k = PRF(master_key, iv)` which is then used to encrypt the
plaintext. (See also: The GenerateKey API `/v1/key/generate/`)

The new API endpoint returns a ciphertext as the encrypted plaintext:
```
{
  "ciphertext": "..."
}
```

The optional context value is cryptographically bound to the ciphertext
such that the same context value must be provided again for decryption.
The context value is not encrypted.

-----

The new encrypt API is not designed to encrypt large data streams.
Instead, it should be used to encrypt/wrap small amounts of data - e.g.
other cryptographic keys or metadata.