vault: allow custom engine paths
This commit allows users to specify a custom
engine path for the K/V secret engine and
AppRole authentication method, and therefore,
fixes #55.

Vault allows engines to be mounted at arbitrary
paths. For instance, you can mount the K/V secret
engine at `/kv` or `/secrets`.
Further, Vault allows multiple instances of the same
engine to be present at the same time. Therefore,
supporting user-defined engine paths is necessary.

The same is true for authentication methods/engines,
like AppRole.

This adds two configuration options `engine: <engine-name>`.
One is for the K/V engine and the other one is for the
AppRole engine.
However, if not specified the server assumes that the
K/V engine is mounted at the "default" `/kv` path and
similarly, that the AppRole engine is mounted at the
"default" `/approle` path.