server: add server startup message
This commit adds a `-q, --quiet` flag to the
`kes server` command.

If not specified the server now prints a startup
message that contains the server endpoint(s),
root identity, key store location a.s.o.

If the `-q` flag is specified the server does
not print a startup message and behaves as before.