vault: unify errors returned to client
This commit changes the errors returned to
the KES client when KES fails to create,
get or delete a master key for some (unexpected)
reason.

Before this commit, the KES server returned errors
that happened between KES and Vault to the client.
This had two downsides:
 - The client could become aware of what KMS/key store
   the KES server is using.
 - The client may encounter Vault-specific error messages.

Now, the KES server returns always the same `Bad Gateway` error
with a generic error message indicating that the communication between
the KES server and its KMS failed for some reason.

This behavior is already implemented by the Gemalto KeySecure
integration.