fix docker file for edge image
This commit fixes the docker file for the `minio/kes:edge`
image. The previous docker file incorrectly assumed that
the CI will build and push a docker image as for the `edge`
label to the docker hub.

Now, the docker file clones the minio/kes repo from scratch and
builds the `kes` binary via the Go 1.14 toolchain.