reject unknown fields in the yaml config file
This commit changes the behavior of config file parsing.
Now, a KES server rejects config files with unknown fields
by making the decoding strict.

The main reason for this change is that when a user specifies
a non-existing `keys` backend or an existing `keys` backend has
a typo the KES server will start with a in-memory backend key store
instead of reporting an error.