adjust CLI to be more like modern unix commands
This commit refactors the CLI help messages
and error reporting to follow modern unix command
styles.

In particular, the command help messages contains
a list of options, an short description and a example
showing the command.

Further, this commit unifies the error messages to start
with: `Error:` to visually distinguish error messages from
command results.