docker: fix versioning in container images
This commit address a problem w.r.t. versioning
for docker containers. Now, the container fetches the
binary with the correct version and for the right
architecture on startup.

Before, the container reported an unexpected version
when executing `docker run <id> --version`.

The main problem was that binary versions were not aligned
with docker image tags. A container running the image
`v0.13.0` should also run the binary `v0.13.0` - not the
latest binary version.