add mechanism to create keys in a declarative way
This commit adds a mechanism to create keys in
a declarative way. Now, keys can be specified
in the `keys` section of the config file:

```
keys:
  - name: my-key
  - name: my-key2
```

The KES server will create these keys before
startup.

This commit is a breaking change. The `keys`
section was used to define the KMS/KeyStore
backend. Now, the KMS backend can be specified
in the `keystore` section.

However, the KES server will support the previous
config file format for a limited amount of time.
It will first try to parse the config file assuming
the current format and - if that fails - try to
parse the config file again assuming the previous
format.