plugin: add generic KES KeyStore plugin v1
This commit adds a generic KeyStore plugin API
to the KES server.

Now, the KES server can talk to an external plugin
that implements four REST endpoints for creating,
listing, accessing and deleting keys.

This commit also adds the plugin specification.

With KeyStore plugins it is possible to support
proprietary/custom KMS/KeyStore implementations
that should not be published.

For a more detailed explanation about plugins
and how they should be implemented take a look
at the generic/spec-v1.md

Signed-off-by: Andreas Auernhammer <aead@mail.de>